BigDataFlink/flink_job/requirements.txt
pyflink
psycopg2-binary
protobuf<5.0.0